requirements.txt
fastapi==0.117.1
uvicorn==0.36.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
jinja2==3.1.6
aiofiles==24.1.0
minio==7.2.16
email-validator==2.3.0
fastapi-mail==1.4.1
python-dotenv==1.0.0
redis==6.4.0
psycopg2-binary==2.9.10
alembic==1.13.1
